refactor(dsg): moving to sepereded folders
