+++ packages/amplication-data-service-generator/src/server/resource/dto/to-many/create-create-nested-many-without-input/index.ts
@@ -1,0 +1,2 @@
+export * from "./create-create-nested-many-without-input";
+export * from "./create-ast-id";
