+++ packages/amplication-data-service-generator/src/server/resource/dto/to-many/create-update-many-without-input/index.ts
@@ -1,2 +1,2 @@
 export * from "./create-update-many-without-input";
-export * from "./create-id";
+export * from "./create-ast-id";
